project completed

+++ src/app/mitgliederliste/mitgliederliste.component.ts
@@ -1,3 +1,4 @@
+import { DialogService } from './../shared/dialog.service';
 import { Component, Input, OnInit } from '@angular/core';
 import { Mitglied } from 'src/model/Interface/mitglied.class';
 import { ApiService } from '../shared/api.service';
@@ -14,7 +15,7 @@ import { DetailansichtComponent } from '../detailansicht/detailansicht.component
 export class MitgliederlisteComponent implements OnInit {
 
 
-  constructor(public api: ApiService) { }
+  constructor(public api: ApiService, private dialogService: DialogService) { }
 
   addedUsers = this.api.addedUsers;
   Liste = this.api.Liste
@@ -32,15 +33,41 @@ export class MitgliederlisteComponent implements OnInit {
   }
   /**
    * This function deletes the current User from List
-   * @param id - User Id from current User
+   * @param {number} id - User Id from current User
    */
   deleteUser(id: any) {
-    this.api.deleteUser(id).subscribe((result: any) => {
 
-      this.Liste.splice(result, 1)
-      console.log(result)
-    })
+
+    this.dialogService.openConfirmDialog('Wollen Sie diesen Eintrag wirklich löschen?').afterClosed().subscribe(res => {
+      if (res) {
+        this.api.deleteUser(id).subscribe((result: any) => {
+
+          this.Liste.splice(result, 1)
+          console.log(result)
+        });
+        this.api.successMessage('Das Mitglied wurde erfolgreich gelöscht!')
+      }
+    });
+
   }
+
+  /**
+   * This function deletes the new created user
+   * @param {number} i - index of current User
+   */
+
+  deleteNewUser(i: any) {
+    i = 0;
+
+    this.dialogService.openConfirmDialog('Wollen Sie diesen Eintrag wirklich löschen?').afterClosed().subscribe(res => {
+      if (res) {
+        this.addedUsers.splice(i, 1);
+        this.api.successMessage('Das Mitglied wurde erfolgreich gelöscht!')
+      }
+    });
+
+  }
+
